Moved the validate functions outside the ui.go
